staticx only for oldest

+++ src/travis/linux-x86_64-install.sh
@@ -5,18 +5,22 @@ export GAMVERSION=`gam/gam version simple`
 cp LICENSE gam
 cp whatsnew.txt gam
 cp GamCommands.txt gam
-GAM_ARCHIVE=gam-$GAMVERSION-$GAMOS-$PLATFORM.tar.xz
+this_glibc_ver=$(ldd --version | awk '/ldd/{print $NF}')
+GAM_ARCHIVE=gam-$GAMVERSION-$GAMOS-$PLATFORM-$this_glibc_ver.tar.xz
 tar cfJ $GAM_ARCHIVE gam/
 echo "PyInstaller GAM  info:"
 du -h gam/gam
 time gam/gam version extended
 
-GAM_LEGACY_ARCHIVE=gam-$GAMVERSION-$GAMOS-$PLATFORM-legacy.tar.xz
-$python -OO -m staticx gam/gam gam/gam-staticx
-strip gam/gam-staticx
-rm gam/gam
-cp staticx-gam.sh gam/gam
-tar cfJ $GAM_LEGACY_ARCHIVE gam/
-echo "Legacy StaticX GAM info:"
-du -h gam/gam-staticx
-time gam/gam version extended
+dist=$(lsb_release --codename --short)
+if [[ "$dist" == "precise" ]]; then
+  GAM_LEGACY_ARCHIVE=gam-$GAMVERSION-$GAMOS-$PLATFORM-legacy.tar.xz
+  $python -OO -m staticx gam/gam gam/gam-staticx
+  strip gam/gam-staticx
+  rm gam/gam
+  cp staticx-gam.sh gam/gam
+  tar cfJ $GAM_LEGACY_ARCHIVE gam/
+  echo "Legacy StaticX GAM info:"
+  du -h gam/gam-staticx
+  time gam/gam version extended
+fi
